fix: move UI-connect methods from SQL module
Fix some issues with their interaction
