+++ app/src/main/kotlin/viewmodel/FastFourierTransform.kt
@@ -1,4 +1,4 @@
-package viewmodel.graph
+package viewmodel
 
 /**
  *  Copyright (C) 2021 José Alexandre Nalon
